fix: change MapManagerData->textureIndices from Float32Array to Uint32Array

+++ src/renderer/texture/MapManagerData.ts
@@ -1,7 +1,6 @@
 export class MapManagerData {
     public static buffer: SharedArrayBuffer = null;
 
-    //todo change to Uint32Array
-    public static textureIndices: Float32Array = null;
+    public static textureIndices: Uint32Array = null;
     public static textureCounts: Uint8Array = null;
 }
